requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.1.5
django-cors-headers==4.2.0
django-picklefield==3.1
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
